+++ src/shadowfax/algorithm_2d/delaunay_2d.h
