+++ src/shadowfax/algorithm_2d/delaunay_functions_2d.h
