+++ src/shadowfax/algorithm_2d/geometry_2d.h
