+++ src/shadowfax/algorithm_2d/triangle.h
