+++ src/shadowfax/algorithm_2d/voronoi_2d.h
