+++ src/shadowfax/algorithm_2d/voronoi_functions_2d.h
