+++ src/shadowfax/algorithm_3d/geometry_3d.h
